refactor(nbody): extract Trinary engine into shared src/lib/nbody library
Move the framework-agnostic gravitational core out of the Trinary app folder
into a first-class shared library (mirrors src/lib/particles):

- physics.ts        -> src/lib/nbody/integrator.ts
- presets.ts        -> src/lib/nbody/scenarios.ts
- analysis/*        -> src/lib/nbody/analysis/*
- new src/lib/nbody/index.ts barrel re-exports the public surface

All Trinary consumers (app + lab, including the Web Worker entry points that
reach the engine transitively) now import from '@/lib/nbody'. The library is
pure TypeScript with no React/Three.js/DOM dependencies, so it bundles cleanly
on the main thread and inside workers. Behaviour is unchanged; npm run build
passes.

Phase 1 of docs/TRINARY_ROADMAP.md.
