+++ src/lib/nbody/analysis/types.ts
