+++ src/lib/nbody/integrator.ts
